INT -> SMALLINT

+++ greyfish-core/src/main/resources/h2/h2_create_database.sql
@@ -47,7 +47,7 @@ CREATE TABLE SIMULATION (
 
 CREATE TABLE AGENT_INTERACTION (
         id INT NOT NULL PRIMARY KEY AUTO_INCREMENT,
-        type_name_id INT NOT NULL,
+        type_name_id SMALLINT NOT NULL,
         source_id INT NOT NULL,
         target_id INT NOT NULL,
         simulation_step INT NOT NULL)
